added footer elements to Footer.jsx
