+++ src/App.jsx
@@ -6,10 +6,11 @@ import Event from "./Components/Event";
 // import Footer from "./Components/Footer";
 // import Header from "./Components/Header";
 import NewEventForm from "./Components/NewEventForm";
+import Attendees from "./Components/Attendees";
 
 function App() {
   
-  const [events, setEvents] = useState(eventsData); // const [events, setEvents] = useState(eventsData);: Initializes events state with eventsData. This state holds the array of event objects.
+  const [events, setEvents] = useState(eventsData); // const [events, setEvents] = useState(eventsData);: Initializes events state with eventsData. This state holds the array of event objects. useState used in several components.
 
   const [showAttendees, setShowAttendees] = useState(false); // This controls whether the attendees of an event are displayed.
 
@@ -25,18 +26,7 @@ function App() {
     people: [],
   });
 
-  function addEvent() {
-    const createEvent = {
-      id: generateUniqueID(),
-      eventType: selectOption,
-      name: newEvent.name,
-      organizer: newEvent.organizer,
-      eventImage: newEvent.eventImage || null,
-      date: newEvent.date,
-      people: [],
-    };
-    handleAddEvent(createEvent);
-  }
+
 
   function handleSelectChange(e) {
     setSelectOption(e.target.value);
@@ -98,16 +88,15 @@ function App() {
       <main>
         <NewEventForm
         handleSubmit={handleSubmit} handleTextChange= {handleTextChange} newEvent = {newEvent} handleSelectChange={handleSelectChange} />
+          <div className="events">
+          {/* <ul> added to Event.jsx */}
         <Event events={events} toggleEventAttendees={toggleEventAttendees} updateEventAttendance={updateEventAttendance} />
+        <Attendees showAttendees={showAttendees} toggleEventAttendees={toggleEventAttendees} updateEventAttendance={updateEventAttendance}/>
+          {/* </ul> */}
+          </div>
       </main>
       <>
-        <footer>
-          <ul>
-            <li>Contact</li>
-            <li>About</li>
-            <li>Legal</li>
-          </ul>
-        </footer>
+        <Footer />
       </>
     </div>
   );
