+++ src/Components/Event.jsx
@@ -1,13 +1,14 @@
-import { useState } from "react"
+
 import Attendees from './Attendees';
-export default function Event({ events, toggleEventAttendees, updateEventAttendance }) {
-  const [showAttendees, setShowAttendees] = useState(false);
+export default function Event({ events,  updateEventAttendance, showAttendees }) {
+  
+
+  
   return (
 
+  <ul>
 
-
-<div className="events">
-          <ul>
+          
             {events.map((event) => {
               const { people: attendees } = event;
 
@@ -21,72 +22,21 @@ export default function Event({ events, toggleEventAttendees, updateEventAttenda
                     <br />
                     <span>Organized by: {event.organizer} </span>
                     <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>
-                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-                      </button>
-
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img
-                                    src={attendee.avatar}
-                                    alt={attendee.firstName}
-                                  />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button
-                                    className="clickable"
-                                    onClick={() =>
-                                      updateEventAttendance(
-                                        event.id,
-                                        attendee.id
-                                      )
-                                    }
-                                  >
-                                    Attending:
-                                  </button>
-                                  <span>
-                                    {attendee.attendance ? "✅" : "❌"}
-                                  </span>
-                                </p>
-
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
+                    <Attendees events={events} showAttendees={showAttendees} updateEventAttendance={updateEventAttendance} attendees={attendees}/>
                   </li>
                 </>
               );
             })}
-          </ul>
-        </div>
+  
+          
+        
+        </ul>
+        
   )
 }
 
 
 // attendees is nested inside events so add html for attendees to Attendees.jsx
 //might have to import useStates or functions
-
-{/* <li key={event.id}> */}
-{/* event details and attendees toggle button */}
-{/* <Attendees  */}
-//   showAttendees={showAttendees} 
-//   attendees={event.people} 
-//   eventId={event.id} 
-//   updateEventAttendance={updateEventAttendance}
-// />
-// </li>
+// <Attendees  showAttendees={showAttendees} attendees={event.people} eventId={event.id} updateEventAttendance={updateEventAttendance} />
+//
